Modularize layout logic
