+++ src/screen/gemx.rs
@@ -44,7 +44,7 @@ pub fn render_gemx<B: Backend>(f: &mut Frame<B>, area: Rect, state: &mut AppStat
     // Ensure we always have valid root nodes before any layout logic
     state.ensure_valid_roots();
     if state.auto_arrange {
-        state.recalculate_roles();
+        crate::layout::roles::recalculate_roles(state);
     }
 
     // Validate again in case role recalculation removed all roots
@@ -144,63 +144,7 @@ pub fn render_gemx<B: Backend>(f: &mut Frame<B>, area: Rect, state: &mut AppStat
 
     use std::collections::HashSet;
     let reachable_ids: HashSet<NodeID> = drawn_at.keys().copied().collect();
-    if state.auto_arrange {
-        let node_ids: Vec<NodeID> = state.nodes.keys().copied().collect();
-        for id in node_ids {
-            if state.fallback_this_frame {
-                continue;
-            }
-            let node = match state.nodes.get(&id) {
-                Some(n) => n,
-                None => continue,
-            };
-            if state.root_nodes.contains(&id)
-                || drawn_at.contains_key(&id)
-                || reachable_ids.contains(&id)
-                || state.fallback_promoted_this_session.contains(&id)
-            {
-                continue;
-            }
-            if node.children.is_empty() {
-                continue;
-            }
-
-            state.root_nodes.push(id);
-            state.root_nodes.sort_unstable();
-            state.root_nodes.dedup();
-            state.fallback_this_frame = true;
-            state.fallback_promoted_this_session.insert(id);
-
-            let Some(n) = state.nodes.get_mut(&id) else {
-                eprintln!("❌ Fallback failed: Node {} not found.", id);
-                return;
-            };
-
-            if n.x == 0 && n.y == 0 {
-                n.x = state.fallback_next_x;
-                n.y = state.fallback_next_y;
-                state.fallback_next_y += 3;
-                if state.fallback_next_y > area.height as i16 - 4 {
-                    state.fallback_next_y = GEMX_HEADER_HEIGHT + 2;
-                    state.fallback_next_x += 20;
-                }
-                if state.debug_input_mode {
-                    eprintln!(
-                        "\u{1F4D0} Placed Node {} at x={}, y={}",
-                        id, n.x, n.y
-                    );
-                }
-            }
-
-            drawn_at.insert(id, Coords { x: n.x, y: n.y });
-            node_roles.insert(id, LayoutRole::Root);
-
-            crate::log_debug!(state, "\u{26a0} Promoted Node {} to root (label-safe)", id);
-
-            break;
-        }
-
-    }
+        crate::layout::fallback::promote_unreachable(state, &reachable_ids, &mut drawn_at, &mut node_roles, area.height as i16);
 
     for (&id, _) in &state.nodes {
         if !drawn_at.contains_key(&id) {
